refactor: introduce engine package
