+++ pkg/engine/functions/strings.go
