+++ pkg/runner/operations/command/operation.go
@@ -10,10 +10,10 @@ import (
 	"github.com/kyverno/chainsaw/pkg/apis/v1alpha1"
 	apibindings "github.com/kyverno/chainsaw/pkg/runner/bindings"
 	"github.com/kyverno/chainsaw/pkg/runner/check"
-	environment "github.com/kyverno/chainsaw/pkg/runner/env"
 	"github.com/kyverno/chainsaw/pkg/runner/logging"
 	"github.com/kyverno/chainsaw/pkg/runner/operations"
 	"github.com/kyverno/chainsaw/pkg/runner/operations/internal"
+	environment "github.com/kyverno/chainsaw/pkg/utils/env"
 	restutils "github.com/kyverno/chainsaw/pkg/utils/rest"
 	"github.com/kyverno/pkg/ext/output/color"
 	"k8s.io/client-go/rest"
